Ensure output files are closed in #unzip
After debugging an issue which I originally thought was related to
Presso, me and @grddev realized that JRuby doesn't close the target file
in IO.copy_stream when passed a path. At least not in 1.7.23. I don't
know under which circumstances this could be an issue, but it should be
better to close it than not to.